LPD-49121 Extract Select's internal component to SingleSelectBase
